Flatten the single-module tiles package into xdas/tiles.py
