+++ docs/release-notes.md
@@ -8,7 +8,7 @@
 ### Breaking Changes
 - Passing a bare read function as `engine` is no longer supported: subclass `xdas.io.Engine` instead (see the data-formats documentation). Combining a configured engine instance with `vtype`, `ctype` or extra engine keywords raises a `ValueError` (@atrabattoni).
 - The miniseed `ctype` argument is now honored: it previously routed to an unused attribute and the reader always built interpolated time coordinates. The default is unchanged (@atrabattoni).
-- **Tile-backed virtual arrays.** The new `xdas.tiles` package (ported from the 0.3 line) exposes multi-file archives as one lazy `TileArray`: positive-step slicing, concatenation (including along a new dimension), and whole-array reductions all stay lazy, and reads touch only the tiles a selection overlaps. The Silixa TDMS and MiniSEED engines now emit tile-backed data arrays, replacing the serialized-dask-graph fallback; Silixa reads gained time-axis push-down (@atrabattoni).
+- **Tile-backed virtual arrays.** The new `xdas.tiles` module (ported from the 0.3 line) exposes multi-file archives as one lazy `TileArray`: positive-step slicing, concatenation (including along a new dimension), and whole-array reductions all stay lazy, and reads touch only the tiles a selection overlaps. The Silixa TDMS and MiniSEED engines now emit tile-backed data arrays, replacing the serialized-dask-graph fallback; Silixa reads gained time-axis push-down (@atrabattoni).
 - Tile-backed data arrays round-trip through the native xdas netCDF format: the tile manifest is stored as a `__tiles__` sibling group (@atrabattoni).
 - **Optional `tiles` vtype for every HDF5 engine.** `open_dataarray`/`open_mfdataarray` with `vtype="tiles"` back the returned array with a lazy `TileArray` instead of an HDF5 virtual source, for the asn, febus, terra15, apsensing, prodml, and native xdas engines. Saved tile views are directly readable by the 0.3 line. Custom engines add tile support by implementing the `load_tile(path, selection, **params)` static half of `xdas.io.Engine` (@atrabattoni).
 - **Febus now defaults to `tiles`.** A tile view describes a Febus file as a single tile — the overlap trimming lives in the reader — whereas the HDF5 backing needs one mapping per block, so its manifest grew with the block count as well as the file count. Every other HDF5 engine still defaults to `hdf5` (@atrabattoni).
